+++ R/get_purity.R
@@ -2,6 +2,8 @@
 #' @param x A chromatogram in matrix format 
 #' @param noise_threshold Threshold to define noise. Highest proportion of maximum absorbance.
 #' @param lambdas Wavelengths to include.
+#' @return Returns indices of retention times where the signal falls below the
+#' specified noise threshold.
 #' @author Ethan Bass
 find_noise <- function(x, noise_threshold=0.01, lambdas){
   lambdas <- which(as.numeric(colnames(x)) %in% lambdas)
@@ -17,6 +19,8 @@ find_noise <- function(x, noise_threshold=0.01, lambdas){
 #' @param noise_threshold Threshold to define noise. Highest proportion of maximum absorbance.
 #' @param lambdas Wavelengths to include
 #' @importFrom stats var
+#' @return Returns the average variance of the signal over the retention times
+#' defined as noise according to \code{\link{find_noise}}.
 #' @author Ethan Bass
 get_noise_variance <- function(x, noise_threshold=.005, lambdas=c(210:400)){
   noise_idx <- find_noise(x = x, noise_threshold = noise_threshold, 
@@ -41,6 +45,12 @@ get_noise_variance <- function(x, noise_threshold=.005, lambdas=c(210:400)){
 #' @param weight scaling parameter affecting stringency of threshold
 #' @param noise_threshold Threshold to define noise. Highest proportion of maximum absorbance.
 #' @param lambdas Wavelengths to include
+#' @return Returns a vector of purity thresholds at each retention time index
+#' within the peak specified by \code{pos}.
+#' @references
+#' Stahl, Mark. “Peak Purity Analysis in HPLC and CE Using Diode-Array Technology.”
+#' Agilent Technologies, April 1, 2003, 16.
+#' /href{https://www.agilent.com/cs/library/applications/5988-8647EN.pdf}
 #' @author Ethan Bass
 get_agilent_threshold <- function(x, pos, weight=1, noise_threshold=.005,
                                   lambdas=c(210:400)){
@@ -57,7 +67,14 @@ get_agilent_threshold <- function(x, pos, weight=1, noise_threshold=.005,
 #' Calculate spectral similarity
 #' @param x A chromatogram in matrix format
 #' @param pos A vector containing peak information
+#' @return Returns a vector of spectral similarities of the reference spectrum
+#' to the spectrum at each timepoint within the peak specified by \code{pos}.
+#' @references
+#' Stahl, Mark. “Peak Purity Analysis in HPLC and CE Using Diode-Array Technology.”
+#' Agilent Technologies, April 1, 2003, 16.
+#' /href{https://www.agilent.com/cs/library/applications/5988-8647EN.pdf}
 #' @author Ethan Bass
+
 get_spectral_similarity <- function(x, pos){
   idx <- seq(as.numeric(pos[2]), as.numeric(pos[3]))
   suppressWarnings(cor(x[as.numeric(pos[1]),], t(x[idx,])))
@@ -71,6 +88,8 @@ get_spectral_similarity <- function(x, pos){
 #' @param x A chromatogram in matrix format
 #' @param pos A vector containing peak information
 #' @param weight weight provided to \code{\link{get_agilent_threshold}}
+#' @return Returns a vector of peak purity values at each timepoint within the
+#' peak specified by \code{pos}.
 #' @references
 #' Stahl, Mark. “Peak Purity Analysis in HPLC and CE Using Diode-Array Technology.”
 #' Agilent Technologies, April 1, 2003, 16.
@@ -86,6 +105,8 @@ get_purity_values <- function(x, pos, weight =1){
 #' @param x A chromatogram in matrix format
 #' @param pos A vector containing peak information
 #' @param cutoff Proportion of maximum absorbance to use as cutoff.
+#' @return Returns indices within the peak specified by \code{pos} with a higher
+#' signal intensity than the specified cutoff.
 #' @author Ethan Bass
 trim_peak <- function(x, pos, cutoff = 0.05){
   idx <- pos[2]:pos[3]
@@ -98,7 +119,14 @@ trim_peak <- function(x, pos, cutoff = 0.05){
 #' @param weight weight provided to \code{\link{get_agilent_threshold}}
 #' @param cutoff Proportion of maximum absorbance to use as cutoff.
 #' Argument to \code{\link{trim_peak}}.
+#' @return Returns the mean purity of the peak specified by \code{pos}, defined
+#' as the proportion of timepoints with purity values below 1.
+#' @references
+#' Stahl, Mark. “Peak Purity Analysis in HPLC and CE Using Diode-Array Technology.”
+#' Agilent Technologies, April 1, 2003, 16.
+#' /href{https://www.agilent.com/cs/library/applications/5988-8647EN.pdf}
 #' @author Ethan Bass
+
 get_mean_purity <- function(x, pos, weight=1, cutoff=0.05){
   p <- get_purity_values(x, pos, weight = weight)
   mean(p[trim_peak(x, pos, cutoff=cutoff)] < 1)